apps/MediaPilot/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pillow
python-dotenv
requests
